refactor(examples): merge demos into a camera-synced 2x2 viewer
Combine insertPointCloud.py and insertPointCloudColor.py into a single
pointcloud_to_octree.py that builds an OcTree and a ColorOcTree from one
scan and shows pointcloud / occupied / occupied (color) / empty in a 2x2
window. The four panes share one trackball, so dragging any pane rotates
all four together.

Move the rendering machinery (the synced widget and the grid window) into
examples/viewer.py so the example reads as a concise walk through the
octomap API.
